Add confirmation dialog and tests

+++ app/src/main/java/dev/lucasnlm/antimine/stats/StatsActivity.kt
@@ -4,11 +4,13 @@ import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import androidx.activity.viewModels
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Observer
 import dagger.hilt.android.AndroidEntryPoint
 import dev.lucasnlm.antimine.R
 import dev.lucasnlm.antimine.common.level.repository.IStatsRepository
+import dev.lucasnlm.antimine.stats.model.StatsModel
 import dev.lucasnlm.antimine.stats.viewmodel.StatsViewModel
 import kotlinx.android.synthetic.main.activity_stats.*
 import kotlinx.coroutines.GlobalScope
@@ -24,25 +26,43 @@ class StatsActivity : AppCompatActivity(R.layout.activity_stats) {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        refreshStats(StatsViewModel.emptyStats)
+
         viewModel.statsObserver.observe(
             this,
             Observer {
-                minesCount.text = it.mines.toString()
-                totalTime.text = formatTime(it.duration)
-                averageTime.text = formatTime(it.averageDuration)
-                totalGames.text = it.totalGames.toString()
-                performance.text = formatPercentage(100.0 * it.victory / it.totalGames)
-                openAreas.text = it.openArea.toString()
-                victory.text = it.victory.toString()
-                defeat.text = (it.totalGames - it.victory).toString()
-
-                invalidateOptionsMenu()
+                refreshStats(it)
             }
         )
 
         GlobalScope.launch {
             viewModel.loadStats()
         }
+    }
+
+    private fun refreshStats(stats: StatsModel) {
+        if (stats.totalGames > 0) {
+            minesCount.text = stats.mines.toString()
+            totalTime.text = formatTime(stats.duration)
+            averageTime.text = formatTime(stats.averageDuration)
+            totalGames.text = stats.totalGames.toString()
+            performance.text = formatPercentage(100.0 * stats.victory / stats.totalGames)
+            openAreas.text = stats.openArea.toString()
+            victory.text = stats.victory.toString()
+            defeat.text = (stats.totalGames - stats.victory).toString()
+        } else {
+            val emptyText = "-"
+            totalGames.text = "0"
+            minesCount.text = emptyText
+            totalTime.text = emptyText
+            averageTime.text = emptyText
+            performance.text = emptyText
+            openAreas.text = emptyText
+            victory.text = emptyText
+            defeat.text = emptyText
+        }
+
+        invalidateOptionsMenu()
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
@@ -56,14 +76,26 @@ class StatsActivity : AppCompatActivity(R.layout.activity_stats) {
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         return if (item.itemId == R.id.delete) {
-            GlobalScope.launch {
-                viewModel.deleteAll()
-            }
+            confirmAndDelete()
             true
         } else {
             super.onOptionsItemSelected(item)
         }
     }
+
+    private fun confirmAndDelete() {
+        AlertDialog.Builder(this)
+            .setTitle(R.string.are_you_sure)
+            .setMessage(R.string.delete_all_message)
+            .setNegativeButton(R.string.cancel, null)
+            .setPositiveButton(R.string.delete_all) { _, _ ->
+                GlobalScope.launch {
+                    viewModel.deleteAll()
+                }
+            }
+            .show()
+    }
+
 
     companion object {
         private fun formatPercentage(value: Double) =
